Removed use-input-mask in favor of plain Inputmask to remove pattern issues
