+++ package.json
@@ -6,7 +6,7 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "@types/inputmask": "^5.0.6",
+    "@types/inputmask": "^5.0.7",
     "@types/jest": "^27.5.2",
     "@types/jstree": "^3.3.43",
     "@types/node": "^16.18.38",
@@ -24,7 +24,6 @@
     "react-helmet": "^6.1.0",
     "react-scripts": "5.0.1",
     "typescript": "^4.9.5",
-    "use-mask-input": "^3.3.6",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
